Switch to Java 11
Deal with memory management issue: Java up to 10 did not work with docker-imposed memory limit, resulting in out of memory errors on the cluster (pod evicted). Java 9+10 and 8u191 had the feature but not on by default, from 11 onward is is on by default.

Also add a memory limit to the call in iati-data-validator to test running with a strict memory limit (see https://docs.docker.com/install/linux/linux-postinstall/#your-kernel-does-not-support-cgroup-swap-limit-capabilities for Ubuntu/Debian)